pyproject.toml
[project]
name = "deep-linear-ts"
version = "0.1.0"
description = "Deep Linear Networks for Time Series: Exploring the power of depth without nonlinearity"
readme = "README.md"
requires-python = ">=3.11"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
license = {text = "MIT"}
keywords = ["deep-learning", "time-series", "linear-networks", "research"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Programming Language :: Python :: 3.11",
]

dependencies = [
    "torch==2.9.0",
    "torchvision==0.24.0",
    "torchaudio==2.9.0",
    "numpy>=1.24.0,<2.0.0",
    "scipy>=1.10.0",
    "scikit-learn>=1.3.0",
    "pandas>=2.0.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "einops>=0.7.0",
    "opt-einsum>=3.3.0",
    "tqdm>=4.65.0",
    "python-dotenv>=1.0.0",
    "rich>=14.2.0",
    "transformers>=4.57.1",
    "datasets<3.0.0",
    "pyarrow<17.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.7.0",
    "ruff>=0.0.280",
    "mypy>=1.4.0",
    "ipython>=8.14.0",
    "jupyter>=1.0.0",
]

[project.scripts]
deep-linear-train = "deep_linear_ts.train:main"
deep-linear-eval = "deep_linear_ts.evaluate:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/deep_linear_ts"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "F", "I", "N", "W"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false

